# backend/src/openapi.yaml
openapi: 3.0.3
info:
  title: 公论 (GongLun) API
  description: 公论结构化讨论系统后端 API
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: 本地开发服务器
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
  /auth/register:
    post:
      tags: [认证]
      summary: 注册新用户
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  minLength: 2
                  maxLength: 30
                password:
                  type: string
                  minLength: 6
            example:
              username: "alice"
              password: "password123"
      responses:
        '201':
          description: 注册成功
        '400':
          description: 参数校验失败
        '409':
          description: 用户名已存在
  /auth/login:
    post:
      tags: [认证]
      summary: 用户登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
            example:
              username: "alice"
              password: "password123"
      responses:
        '200':
          description: 登录成功，返回 JWT
        '401':
          description: 用户名或密码错误
  /auth/logout:
    post:
      tags: [认证]
      summary: 登出
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 已登出
        '401':
          description: 未认证
...
  /topics/{topicId}/messages:
    get:
      tags: [消息]
      summary: 获取话题消息列表
      parameters:
        - in: path
          name: topicId
          required: true
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 消息列表

    post:
      tags: [消息]
      summary: 发布消息
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: topicId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                contentType:
                  type: string
                  enum: [TEXT, MARKDOWN]
                content:
                  type: string
                  maxLength: 20000
                quoteSourceId:
                  type: string
                quotedText:
                  type: string
                  maxLength: 2000
                quoteContextBefore:
                  type: string
                  maxLength: 200
                quoteContextAfter:
                  type: string
                  maxLength: 200
      responses:
        '201':
          description: 消息创建成功
        '403':
          description: 话题已归档
...